+++ src/game/nlcore/common/types.ts
@@ -4,7 +4,7 @@ import { TransformDefinitions } from "@core/elements/transform/type";
 import { IGamePluginRegistry } from "@core/game/plugin/plugin";
 import { LiveGameEventToken } from "@core/types";
 import { GameHistory } from "../action/gameHistory";
-import { GameConfig } from "../gameTypes";
+import { GameConfig, SavedGame } from "../gameTypes";
 
 export * from "@core/elements/type";
 export type {
@@ -12,5 +12,6 @@ export type {
     LiveGameEventToken, Origins,
     ServiceHandlerCtx, TransformDefinitions,
     GameConfig,
+    SavedGame,
 };