+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - Transform state is not updated correctly when the transform is skipped
 - Abort Events are not propagated correctly
 - Incorrect behavior of `router.back`
+- The game state is not flushed correctly
 
 ## [0.4.4] - 2025/5/9